improvement: refoldering zod and API Property Swagger
