+++ libs/entities/src/schemas/auth-schemas/forgot-password/index.swagger.ts
@@ -1,0 +1,7 @@
+import { ApiProperty } from '@nestjs/swagger';
+export class forgotPasswordSwagger {
+  @ApiProperty({
+    example: 'user@example.com',
+  })
+  public email!: string;
+}
