+++ libs/entities/src/schemas/auth-schemas/index.ts
@@ -1,8 +1,8 @@
-export * from './forgotPassword-schema';
-export * from './login-schema';
-export * from './logout-schema';
-export * from './register-schema';
-export * from './req-user-schema';
-export * from './resend-otp-schema';
-export * from './reset-password-schema';
-export * from './verify-schema';
+export * from './forgot-password';
+export * from './login';
+export * from './logout';
+export * from './register';
+export * from './req-user';
+export * from './resend-otp';
+export * from './reset-password';
+export * from './verify';
