+++ libs/entities/src/schemas/auth-schemas/login/index.schema.ts
@@ -1,17 +1,4 @@
-import { ApiProperty } from '@nestjs/swagger';
 import { z } from 'zod';
-
-export class LoginSchema {
-  @ApiProperty({
-    example: '',
-  })
-  public email!: string;
-
-  @ApiProperty({
-    example: 'min length 6, upper case 1, numbers 1',
-  })
-  public password!: string;
-}
 
 export const LoginZodSchema = z.object({
   email: z
