+++ libs/entities/src/schemas/auth-schemas/login/index.swagger.ts
@@ -1,0 +1,13 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class LoginSwagger {
+  @ApiProperty({
+    example: '',
+  })
+  public email!: string;
+
+  @ApiProperty({
+    example: 'min length 6, upper case 1, numbers 1',
+  })
+  public password!: string;
+}
