+++ libs/entities/src/schemas/auth-schemas/logout/index.ts
@@ -1,0 +1,2 @@
+export * from './index.schema';
+export * from './index.swagger';
