+++ libs/entities/src/schemas/auth-schemas/register/index.schema.ts
@@ -1,30 +1,4 @@
-import { ApiProperty } from '@nestjs/swagger';
 import { z } from 'zod';
-
-export class RegisterSchema {
-  @ApiProperty()
-  public fullname!: string;
-
-  @ApiProperty({
-    description: 'Email tidak valid',
-    type: 'string',
-    format: 'email',
-  })
-  public email!: string;
-
-  @ApiProperty({
-    description: 'Password harus lebih dari 6 karakter',
-    minLength: 6,
-    type: 'string',
-  })
-  public password!: string;
-
-  @ApiProperty()
-  public phone_number!: string;
-
-  @ApiProperty()
-  public role_id!: number;
-}
 
 export const RegisterZodSchema = z.object({
   fullname: z.string().nonempty({
