+++ libs/entities/src/schemas/auth-schemas/req-user/index.swagger.ts
@@ -1,0 +1,13 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class ReqUserSwagger {
+  @ApiProperty({
+    example: '',
+  })
+  public email!: string;
+
+  @ApiProperty({
+    example: '41037006280',
+  })
+  public nik!: string;
+}
