+++ libs/entities/src/schemas/auth-schemas/req-user/index.ts
@@ -1,0 +1,2 @@
+export * from './index.schema';
+export * from './index.swagger';
