+++ libs/entities/src/schemas/auth-schemas/resend-otp/index.schema.ts
@@ -1,12 +1,4 @@
-import { ApiProperty } from '@nestjs/swagger';
 import { z } from 'zod';
-
-export class resendOtpSchema {
-  @ApiProperty({
-    example: '',
-  })
-  public email!: string;
-}
 
 export const resendOtpZodSchema = z.object({
   email: z
