+++ libs/entities/src/schemas/auth-schemas/resend-otp/index.swagger.ts
@@ -1,0 +1,8 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class resendOtpSwagger {
+  @ApiProperty({
+    example: '',
+  })
+  public email!: string;
+}
