+++ libs/entities/src/schemas/auth-schemas/reset-password/index.schema.ts
@@ -1,18 +1,4 @@
-import { ApiProperty } from '@nestjs/swagger';
 import { z } from 'zod';
-
-export class newPasswordSchema {
-  @ApiProperty({
-    example: '',
-  })
-  public email!: string;
-
-  @ApiProperty({
-    example: 'min length 6, upper case 1, numbers 1',
-  })
-  public password!: string;
-}
-
 export const newPasswordZodSchema = z.object({
   email: z.string().email({
     message: 'Email tidak valid',
