+++ libs/entities/src/schemas/auth-schemas/reset-password/index.swagger.ts
@@ -1,0 +1,13 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class newPasswordSwagger {
+  @ApiProperty({
+    example: '',
+  })
+  public email!: string;
+
+  @ApiProperty({
+    example: 'min length 6, upper case 1, numbers 1',
+  })
+  public password!: string;
+}
