+++ libs/entities/src/schemas/auth-schemas/reset-password/index.ts
@@ -1,0 +1,2 @@
+export * from './index.schema';
+export * from './index.swagger';
