+++ libs/entities/src/schemas/auth-schemas/verify/index.swagger.ts
@@ -1,0 +1,12 @@
+import { ApiProperty } from '@nestjs/swagger';
+export class verifyOtpSwagger {
+  @ApiProperty({
+    example: '',
+  })
+  public email!: string;
+
+  @ApiProperty({
+    example: 'length 6',
+  })
+  public otp!: string;
+}
