+++ libs/entities/src/schemas/auth-schemas/verify/index.ts
@@ -1,0 +1,2 @@
+export * from './index.schema';
+export * from './index.swagger';
