+++ libs/entities/src/schemas/student-schemas/create-student/index.ts
@@ -1,0 +1,2 @@
+export * from './index.schema';
+export * from './index.swagger';
