+++ libs/entities/src/schemas/student-schemas/update-student/index.ts
@@ -1,0 +1,2 @@
+export * from './index.schema';
+export * from './index.swagger';
