+++ libs/entities/src/schemas/user-schemas/create-user/index.schema.ts
@@ -1,29 +1,6 @@
-import { ApiProperty } from '@nestjs/swagger';
 import { z } from 'zod';
 
-export class UpdateUserSchema {
-  @ApiProperty()
-  public email!: string;
-
-  @ApiProperty()
-  public nik!: string;
-
-  @ApiProperty()
-  public fullname!: string;
-
-  @ApiProperty({
-    example: 'min length 6, upper case 1, numbers 1',
-  })
-  public password!: string;
-
-  @ApiProperty()
-  public role_id!: number;
-
-  @ApiProperty()
-  public photo!: string;
-}
-
-export const UpdateUserZodSchema = z.object({
+export const CreateUserZodSchema = z.object({
   email: z
     .string()
     .email({
@@ -47,8 +24,6 @@ export const UpdateUserZodSchema = z.object({
       message:
         'Password harus memiliki setidaknya 6 karakter dan mengandung setidaknya 1 huruf kecil, 1 huruf besar, dan 1 angka. Tidak boleh mengandung simbol ',
     }),
-  role_id: z.number().optional(),
-  photo: z.string().optional(),
 });
 
-export type TUpdateUserSchema = z.infer<typeof UpdateUserZodSchema>;
+export type TCreateUserSchema = z.infer<typeof CreateUserZodSchema>;
