+++ libs/entities/src/schemas/user-schemas/create-user/index.swagger.ts
@@ -1,0 +1,17 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class CreateUserSwagger {
+  @ApiProperty()
+  public email!: string;
+
+  @ApiProperty()
+  public nik!: string;
+
+  @ApiProperty()
+  public fullname!: string;
+
+  @ApiProperty({
+    example: 'min length 6, upper case 1, numbers 1',
+  })
+  public password!: string;
+}
