+++ libs/entities/src/schemas/user-schemas/create-user/index.ts
@@ -1,0 +1,2 @@
+export * from './index.schema';
+export * from './index.swagger';
