+++ libs/entities/src/schemas/user-schemas/update-user/index.schema.ts
@@ -1,23 +1,6 @@
-import { ApiProperty } from '@nestjs/swagger';
 import { z } from 'zod';
 
-export class CreateUserSchema {
-  @ApiProperty()
-  public email!: string;
-
-  @ApiProperty()
-  public nik!: string;
-
-  @ApiProperty()
-  public fullname!: string;
-
-  @ApiProperty({
-    example: 'min length 6, upper case 1, numbers 1',
-  })
-  public password!: string;
-}
-
-export const CreateUserZodSchema = z.object({
+export const UpdateUserZodSchema = z.object({
   email: z
     .string()
     .email({
@@ -41,6 +24,8 @@ export const CreateUserZodSchema = z.object({
       message:
         'Password harus memiliki setidaknya 6 karakter dan mengandung setidaknya 1 huruf kecil, 1 huruf besar, dan 1 angka. Tidak boleh mengandung simbol ',
     }),
+  role_id: z.number().optional(),
+  photo: z.string().optional(),
 });
 
-export type TCreateUserSchema = z.infer<typeof CreateUserZodSchema>;
+export type TUpdateUserSchema = z.infer<typeof UpdateUserZodSchema>;
