+++ libs/entities/tsconfig.lib.json
