+++ package.json
@@ -39,7 +39,7 @@
     "@nestjs/microservices": "^10.0.5",
     "@nestjs/passport": "^10.0.0",
     "@nestjs/platform-express": "^10.1.2",
-    "@nestjs/swagger": "^7.1.1",
+    "@nestjs/swagger": "^7.1.2",
     "@prisma/client": "5",
     "@swc/helpers": "~0.5.0",
     "@tailwindcss/forms": "^0.5.3",
@@ -93,6 +93,7 @@
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.8.0",
     "styled-components": "^6.0.5",
+    "swagger": "^0.7.5",
     "tozod": "^3.0.0",
     "ts-deepmerge": "^6.2.0",
     "ts-pattern": "^5.0.1",
